# docs/swagger.json
{
    "components": {
        "schemas": {
            "main.Category": {
                "properties": {
                    "deskripsi": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "nama": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "main.Product": {
                "properties": {
                    "harga": {
                        "type": "integer"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "nama": {
                        "type": "string"
                    },
                    "stok": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
    "info": {
        "description": "REST API untuk mengelola produk dan kategori",
        "title": "Kasir API",
        "version": "1.0"
    },
    "externalDocs": {
        "description": "",
        "url": ""
    },
    "paths": {
        "/": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Root endpoint",
                "tags": [
                    "Root"
                ]
            }
        },
        "/api/categories": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/main.Category"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Get all categories",
                "tags": [
                    "Categories"
                ]
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/main.Category",
                                "summary": "category",
                                "description": "Category data"
                            }
                        }
                    },
                    "description": "Category data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.Category"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    }
                },
                "summary": "Create category",
                "tags": [
                    "Categories"
                ]
            }
        },
        "/api/categories/{id}": {
            "delete": {
                "parameters": [
                    {
                        "description": "Category ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Delete category",
                "tags": [
                    "Categories"
                ]
            },
            "get": {
                "parameters": [
                    {
                        "description": "Category ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.Category"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Get category by ID",
                "tags": [
                    "Categories"
                ]
            },
            "put": {
                "parameters": [
                    {
                        "description": "Category ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/main.Category",
                                "summary": "category",
                                "description": "Category data"
                            }
                        }
                    },
                    "description": "Category data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.Category"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Update category",
                "tags": [
                    "Categories"
                ]
            }
        },
        "/api/products": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/main.Product"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Get all products",
                "tags": [
                    "Products"
                ]
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/main.Product",
                                "summary": "product",
                                "description": "Product data"
                            }
                        }
                    },
                    "description": "Product data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.Product"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    }
                },
                "summary": "Create product",
                "tags": [
                    "Products"
                ]
            }
        },
        "/api/products/{id}": {
            "delete": {
                "parameters": [
                    {
                        "description": "Product ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Delete product",
                "tags": [
                    "Products"
                ]
            },
            "get": {
                "parameters": [
                    {
                        "description": "Product ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.Product"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Get product by ID",
                "tags": [
                    "Products"
                ]
            },
            "put": {
                "parameters": [
                    {
                        "description": "Product ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/main.Product",
                                "summary": "product",
                                "description": "Product data"
                            }
                        }
                    },
                    "description": "Product data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.Product"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Update product",
                "tags": [
                    "Products"
                ]
            }
        },
        "/health": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Health check",
                "tags": [
                    "Health"
                ]
            }
        }
    },
    "openapi": "3.1.0"
}